pyproject.toml
[project]
name = "tele-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.21.0",
    "pypdf>=4.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.32.0",
]
